python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apx-agent"
version = "0.16.0"
description = "Standalone agent runtime for Databricks Apps — MCP, A2A, /invocations, tool routing"
readme = "../README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "databricks-sdk>=0.74.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "databricks-openai>=0.1.0",
    "openai-agents>=0.13.0",
]

[project.optional-dependencies]
eval = ["mlflow>=2.0"]
all = ["apx-agent[eval]"]

[dependency-groups]
dev = [
    "uvicorn[standard]>=0.34.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pyright>=1.1.380",
]

[tool.hatch.build.targets.wheel]
packages = ["src/apx_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
include = ["src/apx_agent"]
exclude = ["src/apx_agent/_ui_*.py", "src/apx_agent/_dev.py"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingImports = "warning"

[tool.apx.agent]
name = "apx-agent"
description = "Example agent"
model = "databricks-meta-llama-3-3-70b-instruct"
